基本运行正常，待进一步调试

+++ app/src/main/java/hust/cc/asynchronousacousticlocalization/processing/Algorithm.java
@@ -89,7 +89,7 @@ public class Algorithm {
     }
 
     public static float meanValue(float[] s, int low, int high){
-        long sum = 0;
+        float sum = 0;
         for(int i = low ; i < high ; i++){
             sum += s[(i+s.length)%s.length];
         }
